Add grade calculation functionality to gradebook app
The latest commit comprises the addition of two main functions to the gradebook app, namely `getAverage` and `getGrade`. The `getAverage` function computes the average score from an array of scores. Meanwhile, the `getGrade` function determines the grade level based on the computed average score. This represents the main functionality in our gradebook app project.