Update price value and modify changeDue function
The value for variable "price" has been updated. In the changeDue function, an operation has been added to compute the remaining cash after the purchase, the CID array is updated to reflect the remaining coins, and an HTML paragraph is added in case of cash depletion. The function argument has also been modified for testing.