Add random background color changer feature
The commit includes the creation of the JavaScript, HTML, and CSS files for a new feature - a random background color changer. The JavaScript file contains the logic for generating random indices to be used with an array of color hex values, whereas the HTML and CSS files outline the structure and design of this feature.